requirements.txt
requests==2.32.5
terminaltables==3.1.10
python-dotenv==1.2.1